refactor: use common _resolve_output_files
