+++ graphrag/cli/evaluate.py
@@ -18,6 +18,7 @@ from graphrag.index.create_pipeline_config import create_pipeline_config
 from graphrag.logger.print_progress import PrintProgressLogger
 from graphrag.storage.factory import StorageFactory
 from graphrag.utils.storage import load_table_from_storage
+from graphrag.utils.cli import _resolve_output_files
 
 logger = PrintProgressLogger("")
 
@@ -106,37 +107,3 @@ def _run_evaluate(
 
     logger.success(f"Evaluate Response: \n{response}")
     return response, context_data
-
-def _resolve_output_files(
-    config: GraphRagConfig,
-    output_list: list[str],
-    optional_list: list[str] | None = None,
-) -> dict[str, pd.DataFrame]:
-    """Read indexing output files to a dataframe dict."""
-    dataframe_dict = {}
-    pipeline_config = create_pipeline_config(config)
-    storage_config = pipeline_config.storage.model_dump()  # type: ignore
-    storage_obj = StorageFactory().create_storage(
-        storage_type=storage_config["type"], kwargs=storage_config
-    )
-    for output_file in output_list:
-        df_key = output_file.split(".")[0]
-        df_value = asyncio.run(
-            load_table_from_storage(name=output_file, storage=storage_obj)
-        )
-        dataframe_dict[df_key] = df_value
-
-    # for optional output files, set the dict entry to None instead of erroring out if it does not exist
-    if optional_list:
-        for optional_file in optional_list:
-            file_exists = asyncio.run(storage_obj.has(optional_file))
-            df_key = optional_file.split(".")[0]
-            if file_exists:
-                df_value = asyncio.run(
-                    load_table_from_storage(name=optional_file, storage=storage_obj)
-                )
-                dataframe_dict[df_key] = df_value
-            else:
-                dataframe_dict[df_key] = None
-
-    return dataframe_dict
